requirements.txt
langchain~=0.0.335
streamlit~=1.28.2
tiktoken
python-dotenv~=1.0.0
pinecone-client
pypdf~=3.17.1
joblib~=1.3.2
pandas~=2.1.3
python~=3.10.8